Clear date row before drawing
Closes https://github.com/nwtnni/tock/issues/2.

Because we center the date, a new date that is shorter than the old
one would only overwrite the middle characters.

Clearing the entire row isn't strictly necessary, since we only need
to clear the maximum character count difference between two consecutive
dates. But that sounds like a pain, and the date is written infrequently,
so this should be fine.